requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.2
boto3==1.15.18
botocore==1.18.18
cffi==1.14.3
cryptography==3.1.1
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.4.0
django-livereload-server==0.3.2
django-rest-knox==4.1.0
django-storages==1.10.1
djangorestframework==3.12.1
google-images-download==2.8.0
gunicorn==20.0.4
isort==5.5.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-http-client==3.3.1
pytz==2020.1
s3transfer==0.3.3
selenium==3.141.0
sendgrid==6.4.7
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
toml==0.10.1
tornado==6.0.4
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
